refactor: 재구매 체크박스 컨텍스트로 리팩터링
